requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
bleach==6.1.0
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.15
drf-spectacular==0.27.0
flake8==6.1.0
importlib-resources==6.1.1
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
mccabe==0.7.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
webencodings==0.5.1
whitenoise==6.6.0
zipp==3.17.0
